mgr/cephadm/tests: move with_service to fixtures.py
(cherry picked from commit 50ecd89d790fb33a7dc297ae1eeb7751dd58cba2)
